from Ӽɛʀօռօɨɖ.

+++ Ӽɛʀօռօɨɖ/Ӽɛʀօռօɨɖɦʊɮ/on.py
@@ -13,50 +13,30 @@
             𝐂𝐨𝐩𝐲𝐫𝐢𝐠𝐡𝐭 (𝐂) 𝟐𝟎𝟐𝟏 𝗛𝘆𝗽𝗲𝗩𝗼𝗶𝗱𝗦𝗼𝘂𝗹 | 𝗛𝘆𝗽𝗲𝗩𝗼𝗶𝗱𝗟𝗮𝗯 | 𝗛𝘆𝗽𝗲𝗩𝗼𝗶𝗱𝘀
 |•••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••|        
 ⇜⊷°•♪   🦋 Ӽɛʀօռօɨɖ🦋   ♪•°⊶⇝         |           ⇜⊷°•♪   🦋 Ӽɛʀօռօɨɖ🦋   ♪•°⊶⇝"""
-from ɖօօʍ_ʀօօʍ import *
-from ǟʊȶօ_քʊʀɢɛʀ import *
-from ʟɨɮʀǟʀʏ_ʀօօʍ import *
-from Ӽɛʀօռօɨɖʍʊֆɨƈ import *
-from ƈʊֆȶօʍ_ʄɨʟȶɛʀֆ import *
+from Ӽɛʀօռօɨɖ.ɖօօʍ_ʀօօʍ import *
+from Ӽɛʀօռօɨɖ.ǟʊȶօ_քʊʀɢɛʀ import *
+from Ӽɛʀօռօɨɖ.ʟɨɮʀǟʀʏ_ʀօօʍ import *
+from Ӽɛʀօռօɨɖ.Ӽɛʀօռօɨɖʍʊֆɨƈ import *
+from Ӽɛʀօռօɨɖ.ƈʊֆȶօʍ_ʄɨʟȶɛʀֆ import *
 '|••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••|'
+
 
 
 '|••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••|'
 @Client.on_message(
 xero_basic_fils
 & xero_self_fils
-& xero_xemp_fils
-& filters.command("resume", prefixes=DYNO_COMMANDK))
-async def resume_playing(_, m: Message):
-    xep.xeronoid_voixe.resume_playout()
-    reply = await m.reply_text(f"{emoji.PLAY_OR_PAUSE_BUTTON} resumed",
-                               quote=False)
-    if xep.xemsg.get('pause') is not None:
-        await xep.xemsg['pause'].delete()
-    await m.delete()
-    await xeronoid_resume_purge((reply,), CLEAN_REMOVER)
-'|••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••|'
-
-
-
-"Below code is for the XeronoidBot only and will be used for logging purposes also"
-'|••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••|'
-@Client.on_message(
-xero_bot_fils
-& xero_self_fils
-& xero_xemp_fils
-& filters.command("resume", prefixes=DYNO_COMMANDK))
-async def pause_playing(client, m: Message):
-    xep.xeronoid_voixe.pause_playout()
-    await xep.update_start_time(reset=True)
+& filters.command("on", prefixes=DYNO_COMMANDK))
+async def join_group_call(client, m: Message):
     xeronoid_voixe = xep.xeronoid_voixe
-    xeronoid_chat_verify = int("•100" + str(xeronoid_voixe.full_chat.id))
-    chat = await client.get_chat(xeronoid_chat_verify)
-    reply = await client.send_animation(
-    animation=xerolink,
-    duration=10,
-    chat_id=LOGGER_ID,
-    caption=f"{XEXO}The Userbot has paused itself in the voice chat of • **{chat.title}**"
-    )
-    xep.xemsg['pause'] = reply
-    await m.delete()
+    if not xeronoid_voixe:
+        xep.xeronoid_voixe = GroupCallFactory(client).get_file_group_call()
+        xep.xeronoid_voixe.add_handler(network_status_changed_handler,
+                                  GroupCallFileAction.NETWORK_STATUS_CHANGED)
+        xep.xeronoid_voixe.add_handler(playout_ended_handler,
+                                  GroupCallFileAction.PLAYOUT_ENDED)
+        await xep.xeronoid_voixe.start(m.chat.id)
+        await m.delete()
+    if xeronoid_voixe and xeronoid_voixe.is_connected:
+        await m.reply_text(f"{emoji.ROBOT} already joined a voice chat")
+'|••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••••|'
